add api to get latest order for create chapter comic/story

+++ Mcsg.Comic.Api/Services/PostService.cs
@@ -1671,8 +1671,7 @@ public partial class PostService : IPostService
         if (chapterPostReq.IsAutoGenerateOrder)
         {
             var maxOrder = (await reader.ReadAsync<float>(false)).FirstOrDefault();
-            newOrder = maxOrder + 1;
-
+            newOrder = (int)maxOrder + 1;
         }
         else
         {
